pyproject.toml
[tool.poetry]
name = "mascai-page-loader"
version = "0.1.0"
description = ""
authors = ["mascai <msumoskalenko@gmail.com>"]
packages = [
  { include = "page_loader" },
]

[tool.poetry.dependencies]
python = "^3.5"
pytest-cov = "^2.8"


[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"

[tool.poetry.scripts]
page-loader = "page_loader.scripts.page_loader:main"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

